24.01.2021: We haven't succeded to create a connection with ESP8266.
the interrupt is not accessed.
(We accessed it by mistake after tuching the pins)
What we tried and didn't work:
1. We tried clearing USART1-SR (TC - transmit complete) by reseting it's
value
This might be the cause of the interrupt - although it is not likely
because we did not enable the TC interrupt.
2. input_floating instead os pull-up
3. waiting
4. Enable RCC for Alternate funcion for PINs - Highly unlikely... we
don't think it should be enabled at all.
On another subject - The init_usart1() test prints are printed twice
even though they are supposed to print once.
We are pertty sure that the BRR is set currectly (but it might be worth
the review...)